chore(refactoring): main code
