+++ src/toDataURI.ts
